requirements.txt
google-generativeai==0.8.3
groq==0.11.0
python-dotenv==1.0.1
pandas==2.2.3
requests==2.32.3